final_project/requirements.txt
ipython
pandas
vertexai
google-cloud-aiplatform
google-cloud-bigquery
google-cloud-bigquery-storage
google-cloud-storage
db-dtypes
streamlit
pillow
numpy
scikit-learn
requests
matplotlib